chore: move outputDifference to parent dir
This file is not specific to descriptors and can be used for
fixed script code as well.

Ticket: BTC-1690

TICKET: BTC-1690
