+++ modules/abstract-utxo/src/transaction/outputDifference.ts
